// retailercart/mobile/retailer-ssfa-pwa/src/app/components/model/distributor.ts
import {Injectable} from '@angular/core';

@Injectable()
export class Distributor {

    cmpCode: string;
    distrCode: string;
    distrName: string;
    distrAddr1: string;
    distrAddr2: string;
    distrAddr3: string;
    pinCode: string;
    phoneNo: string;
    mobileNo: string;
    contactPerson: string;
    emailID: string;
    gstStateCode: string;
    dayOff: string;
    distrStatus: string;
    image: string;
    gstStateName: string;
    distributorLOBMapping = [];
}